+++ lib/ModelMixins/CatalogMemberMixin.ts
@@ -10,7 +10,7 @@ import GroupMixin from "./GroupMixin";
 import MappableMixin from "./MappableMixin";
 import ReferenceMixin from "./ReferenceMixin";
 
-type CatalogMember = Model<CatalogMemberTraits>;
+export type CatalogMember = Model<CatalogMemberTraits>;
 
 function CatalogMemberMixin<T extends Constructor<CatalogMember>>(Base: T) {
   abstract class CatalogMemberMixin extends AccessControlMixin(Base) {