+++ lib/Models/CkanCatalogGroup.ts
@@ -11,7 +11,7 @@ import UrlMixin from "../ModelMixins/UrlMixin";
 import CkanCatalogGroupTraits from "../Traits/CkanCatalogGroupTraits";
 import CkanSharedTraits from "../Traits/CkanSharedTraits";
 import ModelReference from "../Traits/ModelReference";
-import CatalogGroup from "./CatalogGroupNew";
+import CatalogGroup from "./CatalogGroup";
 import { CkanDataset, CkanServerResponse } from "./CkanDefinitions";
 import CkanItemReference from "./CkanItemReference";
 import CommonStrata from "./CommonStrata";